Add if statement confirming user is superuser before amending or deleting product in Product views.py

+++ products/views.py
@@ -56,6 +56,9 @@ def product_detail(request, product_id):
 @login_required
 def add_product(request):
     """ Add a product to the site """
+    if not request.user.is_superuser:
+        return redirect(reverse('home'))
+
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -74,6 +77,9 @@ def add_product(request):
 @login_required
 def edit_product(request, product_id):
     """ Edit a product on the site """
+    if not request.user.is_superuser:
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == "POST":
@@ -95,6 +101,9 @@ def edit_product(request, product_id):
 @login_required
 def delete_product(request, product_id):
     """ Delete a product from site """
+    if not request.user.is_superuser:
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     return redirect(reverse('home'))
